Alterações Representante

+++ src/main/java/br/gov/df/dftrans/scie/domain/Curso.java
@@ -72,7 +72,7 @@ public class Curso implements Serializable {
 		final int prime = 31;
 		int result = 1;
 		result = prime * result + ((codEmec == null) ? 0 : codEmec.hashCode());
-		result = prime * result + id;
+		result = prime * result + (int) id;
 		return result;
 	}
